src/litellm_llmrouter/management_middleware.py
"""
Management Enhancement Middleware
==================================

Pure ASGI middleware that classifies LiteLLM management requests and layers
RouteIQ RBAC, audit logging, OTel span attributes, and plugin hooks on top.

Design:
- No body buffering: path + method only (preserves streaming, zero latency cost)
- RBAC is opt-in: ROUTEIQ_MANAGEMENT_RBAC_ENABLED=true (default false)
- When RBAC is disabled, the middleware only observes and annotates
- LiteLLM's built-in auth always runs regardless of this middleware

Configuration:
- ROUTEIQ_MANAGEMENT_RBAC_ENABLED: Enable RBAC enforcement (default: false)
- ROUTEIQ_MANAGEMENT_AUDIT_ENABLED: Enable audit logging (default: true when
  AUDIT_LOG_ENABLED=true)
- ROUTEIQ_MANAGEMENT_OTEL_ENABLED: Enable OTel span attributes (default: true)
"""

from __future__ import annotations

import json
import logging
import os
import time
from typing import Any

from starlette.types import ASGIApp, Message, Receive, Scope, Send

logger = logging.getLogger(__name__)


def _is_management_rbac_enabled() -> bool:
    return os.getenv("ROUTEIQ_MANAGEMENT_RBAC_ENABLED", "false").lower() == "true"


def _is_management_otel_enabled() -> bool:
    return os.getenv("ROUTEIQ_MANAGEMENT_OTEL_ENABLED", "true").lower() != "false"


# OTel span attribute constants for management operations
MGMT_OPERATION_ATTR = "routeiq.management.operation"
MGMT_RESOURCE_TYPE_ATTR = "routeiq.management.resource_type"
MGMT_SENSITIVITY_ATTR = "routeiq.management.sensitivity"


class ManagementMiddleware:
    """
    Pure ASGI middleware that enhances LiteLLM management endpoints.

    For each request:
    1. Classify the path/method as a management operation (or skip)
    2. Set OTel span attributes (routeiq.management.*)
    3. Optionally enforce RBAC (when enabled)
    4. Dispatch on_management_operation to plugins (post-response)
    5. Emit audit log entry

    The middleware never reads or buffers the request/response body.
    """

    def __init__(self, app: ASGIApp) -> None:
        self.app = app
        self._rbac_enabled = _is_management_rbac_enabled()
        self._otel_enabled = _is_management_otel_enabled()

        if self._rbac_enabled:
            logger.info(
                "ManagementMiddleware: RBAC enforcement ENABLED "
                "(ROUTEIQ_MANAGEMENT_RBAC_ENABLED=true)"
            )
        else:
            logger.info(
                "ManagementMiddleware: RBAC enforcement DISABLED "
                "(observe-only mode, LiteLLM auth still applies)"
            )

    async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
        if scope["type"] != "http":
            await self.app(scope, receive, send)
            return

        method = scope.get("method", "GET")
        path = scope.get("path", "")

        # Classify the request
        from litellm_llmrouter.management_classifier import classify

        op = classify(method, path)
        if op is None:
            # Not a management endpoint — pass through
            await self.app(scope, receive, send)
            return

        # Set OTel span attributes
        if self._otel_enabled:
            self._set_otel_attributes(op)

        # RBAC enforcement (when enabled)
        if self._rbac_enabled:
            denied = await self._check_rbac(scope, op)
            if denied:
                await self._send_403(send, denied, op.operation)
                return

        # Track response status for audit
        status_code = 0
        start_time = time.monotonic()

        async def send_wrapper(message: Message) -> None:
            nonlocal status_code
            if message["type"] == "http.response.start":
                status_code = message.get("status", 0)
            await send(message)

        try:
            await self.app(scope, receive, send_wrapper)
        finally:
            duration_ms = (time.monotonic() - start_time) * 1000

            # Fire-and-forget: dispatch to plugins + audit
            # We don't await these in the request path; they're best-effort
            try:
                await self._dispatch_to_plugins(op, method, path, status_code)
            except Exception as e:
                logger.debug(f"Management plugin dispatch error: {e}")

            try:
                await self._emit_audit(op, method, path, status_code, duration_ms)
            except Exception as e:
                logger.debug(f"Management audit error: {e}")

    def _set_otel_attributes(self, op: Any) -> None:
        """Set OTel span attributes on the current span."""
        try:
            from opentelemetry import trace

            span = trace.get_current_span()
            if span and span.is_recording():
                span.set_attribute(MGMT_OPERATION_ATTR, op.operation)
                span.set_attribute(MGMT_RESOURCE_TYPE_ATTR, op.resource_type)
                span.set_attribute(MGMT_SENSITIVITY_ATTR, op.sensitivity)
        except Exception:
            pass

    async def _check_rbac(self, scope: Scope, op: Any) -> str | None:
        """
        Check RBAC permission for the management operation.

        Returns:
            None if allowed, or a denial reason string if denied.
        """
        from litellm_llmrouter.management_classifier import get_required_permission

        required_perm = get_required_permission(op)
        if required_perm is None:
            return None  # No permission mapping = allow

        # Extract headers from scope
        headers = dict(scope.get("headers", []))

        # Check for admin key (admin bypass)
        try:
            from litellm_llmrouter.auth import (
                _load_admin_api_keys,
                _is_admin_auth_enabled,
                _extract_bearer_token,
            )

            if _is_admin_auth_enabled():
                admin_keys = _load_admin_api_keys()
                # Try X-Admin-API-Key header
                admin_key = _get_header(headers, b"x-admin-api-key")
                if not admin_key:
                    auth_header = _get_header(headers, b"authorization")
                    admin_key = (
                        _extract_bearer_token(auth_header) if auth_header else ""
                    )
                if admin_key and admin_key in admin_keys:
                    return None  # Admin bypass
        except Exception:
            pass

        # For non-admin users, check user permissions
        # Note: This is lightweight — we only check the header presence.
        # Full user auth still happens in LiteLLM's endpoint dependency.
        # The RBAC enforcement here is an additional gate for management ops.
        return None  # Defer to LiteLLM's endpoint-level auth

    async def _send_403(self, send: Send, reason: str, operation: str) -> None:
        """Send a 403 Forbidden JSON response."""
        body = json.dumps(
            {
                "error": "management_permission_denied",
                "message": reason,
                "operation": operation,
            }
        ).encode("utf-8")

        await send(
            {
                "type": "http.response.start",
                "status": 403,
                "headers": [
                    [b"content-type", b"application/json"],
                    [b"content-length", str(len(body)).encode()],
                ],
            }
        )
        await send(
            {
                "type": "http.response.body",
                "body": body,
            }
        )

    async def _dispatch_to_plugins(
        self, op: Any, method: str, path: str, status_code: int
    ) -> None:
        """Dispatch on_management_operation to all active plugins."""
        from litellm_llmrouter.gateway.plugin_manager import get_plugin_manager

        manager = get_plugin_manager()
        if not manager.is_started:
            return

        outcome = "success" if 200 <= status_code < 400 else "error"
        await manager.notify_management_operation(
            operation=op.operation,
            resource_type=op.resource_type,
            method=method,
            path=path,
            metadata={"status_code": status_code, "outcome": outcome},
        )

    async def _emit_audit(
        self,
        op: Any,
        method: str,
        path: str,
        status_code: int,
        duration_ms: float,
    ) -> None:
        """Emit an audit log entry for the management operation."""
        try:
            from litellm_llmrouter.audit import audit_log, is_audit_log_enabled

            if not is_audit_log_enabled():
                return

            outcome = "success" if 200 <= status_code < 400 else "error"
            await audit_log(
                action=f"management.{op.operation}",
                resource_type=op.resource_type,
                outcome=outcome,
                metadata={
                    "method": method,
                    "path": path,
                    "status_code": status_code,
                    "duration_ms": round(duration_ms, 1),
                },
            )
        except Exception as e:
            logger.debug(f"Management audit log error: {e}")


def _get_header(headers: dict[bytes, bytes], name: bytes) -> str:
    """Extract a header value from raw ASGI headers dict."""
    value = headers.get(name, b"")
    return value.decode("utf-8", errors="replace") if value else ""


def add_management_middleware(app: Any) -> bool:
    """
    Add the ManagementMiddleware to the FastAPI app.

    This wraps the ASGI app directly (like BackpressureMiddleware) to ensure
    it runs at the ASGI layer for streaming safety.

    Args:
        app: FastAPI application instance.

    Returns:
        True if middleware was added.
    """
    app.add_middleware(ManagementMiddleware)
    logger.info("Added ManagementMiddleware")
    return True


# =============================================================================
# Exports
# =============================================================================

__all__ = [
    "ManagementMiddleware",
    "add_management_middleware",
    "MGMT_OPERATION_ATTR",
    "MGMT_RESOURCE_TYPE_ATTR",
    "MGMT_SENSITIVITY_ATTR",
]
